+++ CHANGELOG.md
@@ -14,6 +14,9 @@ _Unreleased_
 * Not allowing the "address" option on polygonal buildings.
 * Removed the "inactive" button for new non-amenities.
 * Imagery list in Settings was refreshing constantly.
+* Presenting 250 top values for `payment:*` keys.
+* Maxar Premium Imagery is back.
+* OpenStreetMap layer zoom 19 is back.
 * Many, many new translations — thanks folks, and thanks to Weblate.
 
 ## 0.4.0